+++ packages/smart-accounts-kit/src/actions/index.ts
@@ -65,8 +65,11 @@ export { erc7715GetSupportedExecutionPermissionsAction as getSupportedExecutionP
 export { erc7715GetGrantedExecutionPermissionsAction as getGrantedExecutionPermissions } from './erc7715GetGrantedExecutionPermissionsAction';
 
 export {
+  type ExpiryRule,
   type GetSupportedExecutionPermissionsResult,
   type GetGrantedExecutionPermissionsResult,
+  type RedeemerRule,
+  type Rule,
   type SupportedPermissionInfo,
   type PermissionTypes,
   type NativeTokenStreamPermission,